Update patch set 1
Patch Set 1:

Hi Yatin,

This is the only guide to build OpenStack RDO packages available ? https://blogs.rdoproject.org/2016/07/how-to-build-new-openstack-packages/

I'm trying to follow it but the syntax there is not ideal (the commands are all messed up).

Running tox locally works for me, but I would like to try to build it using DLR becuase that seems to be where the error is coming from.

Patch-set: 1
Label: Code-Review=0